+++ apps/docs/src/app/core/component-docs/time/examples/time-no-spinners-example/time-no-spinners-example.component.ts
@@ -1,9 +1,15 @@
 import { Component } from '@angular/core';
-import { FdDate } from '@fundamental-ngx/core/datetime';
+import { DatetimeAdapter, FdDate, FdDatetimeAdapter } from '@fundamental-ngx/core/datetime';
 
 @Component({
     selector: 'fd-time-no-spinners-example',
-    templateUrl: './time-no-spinners-example.component.html'
+    templateUrl: './time-no-spinners-example.component.html',
+    providers: [
+        {
+            provide: DatetimeAdapter,
+            useClass: FdDatetimeAdapter
+        }
+    ]
 })
 export class TimeNoSpinnersExampleComponent {
     time = new FdDate().setTime(9, 0, 0);