+++ apps/docs/src/app/platform/component-docs/platform-wizard-generator/platform-wizard-generator-header/platform-wizard-generator-header.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fdp-platform-wizard-generator-header',
+    templateUrl: './platform-wizard-generator-header.component.html'
+})
+export class PlatformWizardGeneratorHeaderComponent {
+    constructor() { }
+}