+++ libs/core/src/lib/carousel/carousel.service.ts
@@ -32,17 +32,23 @@ export interface PanEndOutput {
     providedIn: 'root'
 })
 export class CarouselService implements OnDestroy {
+    /** Event thrown when element is dragged. Emits "true" when drag starts and "false" when drag ends. */
+    readonly dragStateChange: EventEmitter<boolean> = new EventEmitter<boolean>();
+
+    /** Event thrown, when active element is changed */
+    readonly activeChange: EventEmitter<PanEndOutput> = new EventEmitter<PanEndOutput>();
+
     /** Configuration for carousel */
     config: CarouselConfig;
 
     /** Initial active item of carousel, position first + offset */
     active: CarouselItemInterface;
 
-    /** Event thrown when element is dragged. Emits "true" when drag starts and "false" when drag ends. */
-    readonly dragStateChange: EventEmitter<boolean> = new EventEmitter<boolean>();
+    /** Set to true for rtl mode */
+    isRtl = false;
 
-    /** Event thrown, when active element is changed */
-    readonly activeChange: EventEmitter<PanEndOutput> = new EventEmitter<PanEndOutput>();
+    /** carousel items query list */
+    items: QueryList<CarouselItemInterface>;
 
     /** @hidden */
     private _hammer: HammerManager = null;
@@ -59,17 +65,18 @@ export class CarouselService implements OnDestroy {
     /** @hidden */
     private _currentTransitionPx = 0;
 
+    /** return current transition value in px */
     get currentTransitionPx(): number {
         return this._currentTransitionPx;
     }
 
+    /** set current transition value in px */
     set currentTransitionPx(currentTransitionPx: number) {
         this._currentTransitionPx = currentTransitionPx;
     }
 
+    /** @hidden */
     private _elementRef: ElementRef;
-
-    items: QueryList<CarouselItemInterface>;
 
     /** @hidden */
     ngOnDestroy(): void {
@@ -78,6 +85,7 @@ export class CarouselService implements OnDestroy {
         }
     }
 
+    /** set initial values for the service */
     initialise(
         config: CarouselConfig,
         items: QueryList<CarouselItemInterface>,
@@ -92,14 +100,17 @@ export class CarouselService implements OnDestroy {
         }
     }
 
+    /** update config value for the service */
     updateConfig(config: CarouselConfig): void {
         this.config = config;
     }
 
+    /** set element ref */
     setElementRef(elementRef: ElementRef): void {
         this._elementRef = elementRef;
     }
 
+    /** set items and assign active item, if not already there */
     setItems(items: QueryList<CarouselItemInterface>): void {
         this.items = items;
         const carouselArray: CarouselItemInterface[] = this.items.toArray();
@@ -109,7 +120,7 @@ export class CarouselService implements OnDestroy {
     }
 
     /** Change active element */
-    goToItem(item: CarouselItemInterface, smooth?: boolean, languageDirection: string = 'ltr'): void {
+    goToItem(item: CarouselItemInterface, smooth?: boolean): void {
         let index: number = this.getIndexOfItem(item);
         if (this.config.infinite) {
             this._centerActive(index);
@@ -117,12 +128,13 @@ export class CarouselService implements OnDestroy {
             index = this.getIndexOfItem(item);
         }
 
-        this._transitionToIndex(index, smooth, languageDirection);
+        this._transitionToIndex(index, smooth);
 
         this._previousActiveItem = item;
     }
 
-    pickNext(languageDirection = 'ltr'): void {
+    /** pick next element and set it active */
+    pickNext(): void {
         const carouselArray: CarouselItemInterface[] = this.items.toArray();
         if (!this.active) {
             this.active = carouselArray[0];
@@ -130,18 +142,26 @@ export class CarouselService implements OnDestroy {
         const activeItemIndex: number = carouselArray.findIndex((item) => item === this.active);
 
         const itemToActivate = carouselArray[activeItemIndex + 1];
-        this.goToItem(itemToActivate, true, languageDirection);
+        this.goToItem(itemToActivate, true);
         this.active = itemToActivate;
     }
 
-    pickPrevious(languageDirection = 'ltr'): void {
+    /** pick previous element and set it active */
+    pickPrevious(): void {
         const carouselArray: CarouselItemInterface[] = this.items.toArray();
         if (!this.active) {
             this.active = carouselArray[2];
         }
         const activeItemIndex: number = carouselArray.findIndex((item) => item === this.active);
-        const itemToActivate = carouselArray[activeItemIndex - 1];
-        this.goToItem(itemToActivate, true, languageDirection);
+        let itemToActivate = carouselArray[activeItemIndex - 1];
+
+        // case where slides can move infinitely in one direction
+        // handle case where on-load activeItemIndex = 0 and activeItemIndex - 1 returns undefined item from item array.
+        if (!itemToActivate) {
+            itemToActivate = carouselArray[this.items.toArray().length - 1];
+        }
+
+        this.goToItem(itemToActivate, true);
         this.active = itemToActivate;
     }
 
@@ -175,7 +195,7 @@ export class CarouselService implements OnDestroy {
     }
 
     /** @hidden */
-    private _transitionToIndex(index: number, smooth?: boolean, languageDirection?: string): void {
+    private _transitionToIndex(index: number, smooth?: boolean): void {
         let transitionPx: number = this._getSize(this.items.first) * index;
 
         if (smooth) {
@@ -184,16 +204,21 @@ export class CarouselService implements OnDestroy {
             this._elementRef.nativeElement.style.transitionDuration = '0s';
         }
 
-        if (languageDirection === 'ltr' || (this.config.vertical && languageDirection === 'rtl')) {
+        if (!this.isRtl || (this.config.vertical && this.isRtl)) {
             transitionPx = -transitionPx;
         }
         this._transitionCarousel(transitionPx);
     }
 
-    /** Get closes element, based on current transition */
+    /** @hidden Get closes element, based on current transition */
     private _getClosest(): CarouselItemInterface {
-        /** If transition is positive, it'should go to first element */
-        if (!this.config.infinite && this._currentTransitionPx > 0) {
+        /** If transition is positive, it should go to first element */
+        if (
+            !this.config.infinite &&
+            ((!this.config.vertical &&
+                ((!this.isRtl && this._currentTransitionPx > 0) || (this.isRtl && this._currentTransitionPx < 0))) ||
+                (this.config.vertical && this._currentTransitionPx > 0))
+        ) {
             return this.items.first;
         }
 
@@ -206,7 +231,7 @@ export class CarouselService implements OnDestroy {
          * every element should have same width, otherwise it should be looped through all elements,
          * which is not good for performance
          */
-        let index: number = Math.abs(Math.ceil(this._currentTransitionPx / size));
+        let index: number = Math.ceil(Math.abs(this._currentTransitionPx / size));
 
         // When elementsAtOnce > 1, swiping should stop at last index - elementsAtOnce
         if (!this.config.infinite && this.config.elementsAtOnce > 1) {
@@ -215,7 +240,7 @@ export class CarouselService implements OnDestroy {
             }
         }
 
-        index = index + (halfApproached ? 1 : 0);
+        index = index + (halfApproached ? 0 : -1);
         /** Checking if transition went out of scope of array */
         if (this.items.toArray()[index]) {
             return this.items.toArray()[index];
@@ -229,7 +254,7 @@ export class CarouselService implements OnDestroy {
         return this.items.toArray().findIndex((_item) => _item === item);
     }
 
-    /** Getting size of carousel, width for horizontal, height for vertical */
+    /** @hidden Getting size of carousel, width for horizontal, height for vertical */
     private _getSize(item: CarouselItemInterface): number {
         if (this.config.vertical) {
             return item.getHeight();
@@ -274,7 +299,7 @@ export class CarouselService implements OnDestroy {
     }
 
     /**
-     * Animates the carousel to the currently selected slide.
+     * @hidden Animates the carousel to the currently selected slide.
      */
     private _transitionCarousel(transitionPx: number): void {
         this._currentTransitionPx = transitionPx;
@@ -286,7 +311,7 @@ export class CarouselService implements OnDestroy {
         }
     }
 
-    /** Pam Start handler, removes transition duration, */
+    /** @hidden Pam Start handler, removes transition duration, */
     private _handlePanStart(): void {
         this._elementRef.nativeElement.style.transitionDuration = '0s';
         this._prePanTransitionPx = this._currentTransitionPx;
@@ -295,7 +320,6 @@ export class CarouselService implements OnDestroy {
 
     /** @hidden */
     private _hammerSetup(): void {
-
         this._hammer = new Hammer(this._elementRef.nativeElement);
 
         if (this.config.vertical) {