+++ libs/core/src/lib/wizard/wizard.component.ts
@@ -6,6 +6,7 @@ import {
     ContentChild,
     ContentChildren,
     ElementRef,
+    HostBinding,
     HostListener,
     Input,
     OnDestroy,
@@ -68,6 +69,13 @@ export class WizardComponent implements AfterViewInit, OnDestroy {
     @Input()
     contentHeight: string;
 
+    /**
+     * Whether or not apply responsive paddings styling.
+     */
+    @Input()
+    @HostBinding('class.fd-wizard--responsive-paddings')
+    responsivePaddings = false;
+
     /** @hidden */
     @ContentChildren(WizardStepComponent, { descendants: true })
     steps: QueryList<WizardStepComponent>;
@@ -102,7 +110,7 @@ export class WizardComponent implements AfterViewInit, OnDestroy {
         private _elRef: ElementRef,
         private readonly _cdRef: ChangeDetectorRef,
         @Optional() private _dialogBodyComponent: DialogBodyComponent
-    ) {}
+    ) { }
 
     /** @hidden */
     @HostListener('window:resize')
@@ -165,6 +173,7 @@ export class WizardComponent implements AfterViewInit, OnDestroy {
         }
         wizardFooterHeight = this._getWizardFooterHeight();
         dialogOffset = this._getDialogOffset();
+
         return shellbarHeight + wizardNavHeight + wizardFooterHeight + dialogOffset;
     }
 
@@ -293,7 +302,8 @@ export class WizardComponent implements AfterViewInit, OnDestroy {
         const templatesLength = this.contentTemplates.length;
         this.contentTemplates = [];
         let _stepId = 0;
-        for (const step of this.steps.toArray()) {
+        const steps = this.steps.toArray();
+        for (const step of steps) {
             if (step.content) {
                 if (!step.isSummary) {
                     step.content.tallContent = false;
@@ -307,11 +317,12 @@ export class WizardComponent implements AfterViewInit, OnDestroy {
              If the step is completed and appendToWizard is true, hide the nextStep button, unless it's the last step,
              or if there's a summary and it is the second to last step
              */
-            if (step.completed && this.appendToWizard && step !== this._getLastNonSummaryStep()) {
+            if (step.completed && this.appendToWizard && step.content?.nextStep && step !== this._getLastNonSummaryStep()) {
                 step.content.nextStep._getElRef().nativeElement.style.display = 'none';
             } else if (
-                (step.completed && this.appendToWizard) ||
-                (step.visited && step.branching && step.status === CURRENT_STEP_STATUS)
+                step.content?.nextStep &&
+                ((step.completed && this.appendToWizard) ||
+                (step.visited && step.branching && step.status === CURRENT_STEP_STATUS))
             ) {
                 step.content.nextStep._getElRef().nativeElement.style.removeProperty('display');
             }
@@ -323,14 +334,23 @@ export class WizardComponent implements AfterViewInit, OnDestroy {
                 if (step.status === CURRENT_STEP_STATUS && (!step.completed || !this.appendToWizard)) {
                     step.content.tallContent = true;
                 }
-                if (!templatesLength || (!this.appendToWizard && step.status === CURRENT_STEP_STATUS)) {
+
+                const isCurrentStep = !this.appendToWizard && step.status === CURRENT_STEP_STATUS;
+
+                if (!templatesLength || isCurrentStep) {
                     this.contentTemplates = [step.content.contentTemplate];
+
+                    // If we found needed step, break the cycle, since it might set wrong step later.
+                    if (isCurrentStep) {
+                        break;
+                    }
                 } else if (this.appendToWizard && !step.isSummary) {
                     this.contentTemplates.push(step.content.contentTemplate);
                 }
             }
         }
-        const lastVisibleTemplate = this.steps.toArray()[this.contentTemplates.length - 1];
+
+        const lastVisibleTemplate = steps[this.contentTemplates.length - 1];
         if (lastVisibleTemplate && lastVisibleTemplate.content) {
             lastVisibleTemplate.content.tallContent = true;
         }
@@ -484,16 +504,19 @@ export class WizardComponent implements AfterViewInit, OnDestroy {
 
     /** @hidden */
     private _shrinkWhileAnyStepIsTooNarrow(): void {
-        this.stackedStepsLeft = [];
-        this.stackedStepsRight = [];
-        this.steps.first.stepIndicator.setStackedItems(this.stackedStepsLeft);
-        this._getLastNonSummaryStep().stepIndicator.setStackedItems(this.stackedStepsRight);
-        this._resetStepClasses();
-        let i = 0;
-        while (this._anyStepIsTooNarrow() && i < this.steps.length - 1) {
-            i++;
-            this._wizardShrinking();
-        }
+        // Add small delay for elements to render appropriately.
+        setTimeout(() => {
+            this.stackedStepsLeft = [];
+            this.stackedStepsRight = [];
+            this.steps.first.stepIndicator.setStackedItems(this.stackedStepsLeft);
+            this._getLastNonSummaryStep().stepIndicator.setStackedItems(this.stackedStepsRight);
+            this._resetStepClasses();
+            let i = 0;
+            while (this._anyStepIsTooNarrow() && i < this.steps.length - 1) {
+                i++;
+                this._wizardShrinking();
+            }
+        }, 10);
     }
 
     /** @hidden */