+++ libs/platform/src/lib/components/form/select/public_api.ts
@@ -1,4 +1,5 @@
-export { FdpSelectData, FdpSelectionChangeEvent } from './commons/base-select';
+export { BaseSelect, FdpSelectData, FdpSelectionChangeEvent } from './commons/base-select';
 export { SelectComponent } from './select/select.component';
 export { OptionComponent } from './option/option.component';
 export { PlatformSelectModule } from './select.module';
+export { SelectConfig } from './select.config';