+++ libs/platform/src/lib/domain/public_api.ts
@@ -2,3 +2,4 @@ export * from './array-data-source';
 export * from './observable-data-source';
 export * from './data-model';
 export * from './data-source';
+export * from './base-data-provider';